Issue_PreDeadline_6
the jupyter notebook for the coding of calculating 1 day prior to the deadline date and adding it to a csv file.  The code uses the csv "GP1 Dataset_ Assignments.csv" and saves it into a csv "GP1 Dataset_Assignments_PreDeadline.csv" which will also be uploaded. This may require changes to the main  code.